backend/requirements.txt
mysql-connector-python
python-dotenv
tk
